Ajout des repères 1.5°  et 2°
